Organize files, fix adt
