+++ frontend/src/main/java/org/abs_models/frontend/analyser/BitVec.java
@@ -1,4 +1,4 @@
-package org.abs_models.frontend.typechecker.nullable;
+package org.abs_models.frontend.analyser;
 
 import java.util.*;
